migrating identitity to samples v2
