+++ sdk/communication/communication-identity/sample.env
@@ -1,7 +1,3 @@
-# Used in most samples. Retrieve these values from a Communication Services resource
+# Used in most samples. Retrieve these values from a Communication Services instance
 # in the Azure Portal.
-COMMUNICATION_CONNECTION_STRING="endpoint=<endpoint>;accessKey=<accessKey>"
-
-# Our tests assume that TEST_MODE is "playback" by default. You can
-# change it to "record" to generate new recordings, or "live" to bypass the recorder entirely.
-# TEST_MODE=playback
+COMMUNICATION_CONNECTION_STRING="endpoint=https://<resource name>.communication.azure.net/;accessKey=<key>"
