--- sdk/communication/communication-identity/samples/.gitignore
@@ -1,1 +1,0 @@
-.npmrc
