--- sdk/communication/communication-identity/samples/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "typescript/dist",
-    "lib": ["DOM", "ES6"]
-  },
-  "include": ["typescript/src/**.ts"],
-  "exclude": ["typescript/*.json", "**/node_modules/", "../node_modules", "../typings"]
-}
