+++ sdk/communication/communication-identity/samples/v1/javascript/issueToken.js
@@ -2,14 +2,14 @@
 // Licensed under the MIT License.
 
 /**
- * Demonstrates how to use the CommunicationIdentityClient to
+ * @summary Demonstrates how to use the CommunicationIdentityClient to
  * issue a new user token.
  */
 
-import { CommunicationIdentityClient } from "@azure/communication-identity";
+const { CommunicationIdentityClient } = require("@azure/communication-identity");
 
 // Load the .env file if it exists
-import * as dotenv from "dotenv";
+const dotenv = require("dotenv");
 dotenv.config();
 
 // You will need to set this environment variables or edit the following values
@@ -17,9 +17,10 @@ const connectionString =
   process.env["COMMUNICATION_CONNECTION_STRING"] || "<communication service connection string>";
 
 export const main = async () => {
-  console.log("\n== Issue Token Typescript Sample ==\n");
+  console.log("\n== Issue Token TypeScript Sample ==\n");
 
   const client = new CommunicationIdentityClient(connectionString);
+  const scopes = ["chat"];
 
   // Create user
   console.log("Creating User");
@@ -27,29 +28,16 @@ export const main = async () => {
   const user = await client.createUser();
 
   console.log(`Created user with id: ${user.communicationUserId}`);
+  console.log("Issuing Token");
 
-  console.log("Issuing Tokens");
+  // Issue token and get token from response
+  const { token } = await client.getToken(user, scopes);
 
-  // Issue tokens
-  const { token: token1 } = await client.getToken(user, ["chat"]);
-  const { token: token2 } = await client.getToken(user, ["voip"]);
-  const { token: token3 } = await client.getToken(user, ["voip"]);
-
-  console.log("Issued tokens:");
-  console.log(token1);
-  console.log(token2);
-  console.log(token3);
-
-  // Revoke tokens
-  console.log("Revoking Tokens");
-
-  await client.revokeTokens(user);
-
-  console.log("Tokens Revoked");
+  console.log(`Issued token: ${token}`);
 };
 
 main().catch((error) => {
-  console.error("Encountered an error while issuing/refreshing token: ");
+  console.error("Encountered an error while issuing token: ");
   console.error("Request: \n", error.request);
   console.error("\nResponse: \n", error.response);
   console.error(error);
