+++ sdk/communication/communication-identity/samples/v1/javascript/package.json
@@ -1,19 +1,20 @@
 {
-  "name": "azure-communication-identity-user-token-samples-js",
-  "version": "0.1.0",
-  "description": "Azure Communication Identity client library samples for JavaScript",
+  "name": "azure-communication-identity-samples-js",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Azure Communication Services client library samples for JavaScript",
   "engine": {
-    "node": ">=8.0.0"
+    "node": ">=12.0.0"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git"
+    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
+    "directory": "sdk/communication/communication-identity"
   },
   "keywords": [
-    "Azure",
-    "Communication",
-    "Node.js",
-    "JavaScript"
+    "azure",
+    "cloud",
+    "communication"
   ],
   "author": "Microsoft Corporation",
   "license": "MIT",
@@ -21,9 +22,8 @@
     "url": "https://github.com/Azure/azure-sdk-for-js/issues"
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/communication/communication-identity",
-  "sideEffects": false,
   "dependencies": {
     "@azure/communication-identity": "latest",
-    "dotenv": "^8.2.0"
+    "dotenv": "latest"
   }
 }
