+++ sdk/communication/communication-identity/samples/v1/typescript/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "azure-communication-identity-user-token-samples-ts",
-  "version": "0.1.0",
-  "description": "Azure Communication Identity client library samples for TypeScript",
+  "name": "azure-communication-identity-samples-ts",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Azure Communication Services client library samples for TypeScript",
   "engine": {
-    "node": ">=8.0.0"
+    "node": ">=12.0.0"
   },
   "scripts": {
     "build": "tsc",
@@ -11,13 +12,13 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git"
+    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
+    "directory": "sdk/communication/communication-identity"
   },
   "keywords": [
-    "Azure",
-    "Communication",
-    "Node.js",
-    "JavaScript"
+    "azure",
+    "cloud",
+    "communication"
   ],
   "author": "Microsoft Corporation",
   "license": "MIT",
@@ -25,14 +26,12 @@
     "url": "https://github.com/Azure/azure-sdk-for-js/issues"
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/communication/communication-identity",
-  "sideEffects": false,
   "dependencies": {
     "@azure/communication-identity": "latest",
-    "dotenv": "^8.2.0"
+    "dotenv": "latest"
   },
   "devDependencies": {
-    "@types/node": "^8.0.0",
-    "rimraf": "^3.0.0",
-    "typescript": "~3.6.4"
+    "typescript": "~4.2.0",
+    "rimraf": "latest"
   }
 }
