+++ sdk/communication/communication-identity/samples/v1/typescript/sample.env
@@ -1,3 +1,3 @@
-# Used in most samples. Retrieve these values from a Communication Service instance
+# Used in most samples. Retrieve these values from a Communication Services instance
 # in the Azure Portal.
 COMMUNICATION_CONNECTION_STRING="endpoint=https://<resource name>.communication.azure.net/;accessKey=<key>"
