+++ sdk/communication/communication-identity/samples/v1/typescript/src/revokeTokens.ts
@@ -2,25 +2,24 @@
 // Licensed under the MIT License.
 
 /**
- * Demonstrates how to use the CommunicationIdentityClient to
- * revoke a user's tokens.
+ * @summary Demonstrates how to use the CommunicationIdentityClient to
+ * issue a new user token.
  */
 
-const { CommunicationIdentityClient } = require("@azure/communication-identity");
+import { CommunicationIdentityClient } from "@azure/communication-identity";
 
 // Load the .env file if it exists
-const dotenv = require("dotenv");
+import * as dotenv from "dotenv";
 dotenv.config();
 
 // You will need to set this environment variables or edit the following values
 const connectionString =
   process.env["COMMUNICATION_CONNECTION_STRING"] || "<communication service connection string>";
 
-async function main() {
-  console.log("\n== Issue Token Javascript Sample ==\n");
+export const main = async () => {
+  console.log("\n== Issue Token Typescript Sample ==\n");
 
   const client = new CommunicationIdentityClient(connectionString);
-  const scopes = ["chat", "voip"];
 
   // Create user
   console.log("Creating User");
@@ -32,9 +31,9 @@ async function main() {
   console.log("Issuing Tokens");
 
   // Issue tokens
-  const { token: token1 } = await client.getToken(user, scopes);
-  const { token: token2 } = await client.getToken(user, scopes);
-  const { token: token3 } = await client.getToken(user, scopes);
+  const { token: token1 } = await client.getToken(user, ["chat"]);
+  const { token: token2 } = await client.getToken(user, ["voip"]);
+  const { token: token3 } = await client.getToken(user, ["voip"]);
 
   console.log("Issued tokens:");
   console.log(token1);
@@ -47,10 +46,10 @@ async function main() {
   await client.revokeTokens(user);
 
   console.log("Tokens Revoked");
-}
+};
 
 main().catch((error) => {
-  console.error("Encountered an error while revoking tokens: ");
+  console.error("Encountered an error while issuing/refreshing token: ");
   console.error("Request: \n", error.request);
   console.error("\nResponse: \n", error.response);
   console.error(error);
