+++ sdk/communication/communication-identity/samples/v1/typescript/tsconfig.json
@@ -1,13 +1,17 @@
 {
   "compilerOptions": {
+    "target": "ES2018",
     "module": "commonjs",
     "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "strict": true,
     "alwaysStrict": true,
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/**.ts"],
-  "exclude": ["node_modules"]
+  "include": [
+    "src/**.ts"
+  ]
 }
