+++ sdk/communication/communication-identity/test.env
@@ -1,0 +1,3 @@
+# Our tests assume that TEST_MODE is "playback" by default. You can
+# change it to "record" to generate new recordings, or "live" to bypass the recorder entirely.
+# TEST_MODE=playback
