+++ sdk/communication/communication-identity/tsconfig.json
@@ -2,16 +2,16 @@
   "extends": "../../../tsconfig.package",
   "compilerOptions": {
     "outDir": "./dist-esm",
-    "declarationDir": "./types"
+    "declarationDir": "./types",
+    "paths": {
+      "@azure/communication-identity": [
+        "./src/index"
+      ]
+    }
   },
-  "exclude": [
-    "node_modules",
-    "types",
-    "temp",
-    "browser",
-    "dist",
-    "dist-esm",
-    "dist-samples",
-    "./samples/**/*.ts"
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts",
+    "samples-dev/**/*.ts"
   ]
 }
